Do not remove remote
In order for girror to support concurrency, removed the `git remote rm` command from girror.
This means that if someone giggled with the remote _manually_, things might not work well. However, since the git repositories
are managed in a private cached maintained by girror and keyed by a _mangled remote url_, i suspect this is not really important and being concurrency-safe is a more imporant feature.